mux/avi: use uint32_t for buffer offset
The values are read from size_t anyway, nothing to do with a file offset.
They are read from 32-bit variables are written as 32-bit variables.

+++ modules/mux/avi.c
@@ -130,11 +130,11 @@ typedef struct
     int i_streams;
     int i_stream_video;
 
-    off_t i_movi_size;
+    uint32_t i_movi_size;
     avi_stream_t stream[100];
 
     avi_idx1_t idx1;
-    off_t i_idx1_size;
+    uint32_t i_idx1_size;
 
 } sout_mux_sys_t;
